requirements.txt
boto3==1.28.19
botocore==1.31.19
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
osbot-aws==0.6.51
packaging==23.1
Pillow==10.0.0
pipdeptree==2.12.0
pluggy==1.2.0
pypdf==3.13.0
PyPDF2==3.0.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
requests-mock==1.11.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
